#49 feat 아파트 운영정보 화면 구현
