+++ app/src/main/res/drawable/button_setting_2.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android" >
 
-    <solid android:color="#B8D7D5" />
+    <solid android:color="@color/white" />
     <padding
         android:left="10dp"
         android:right="10dp"
