+++ app/src/main/res/layout/fragment_operation_first.xml
@@ -1,0 +1,177 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    tools:context=".ui.entiremenu.AptOperationInfo.OperationFirstFragment">
+
+    <com.google.android.material.appbar.MaterialToolbar
+        android:id="@+id/toolbarAptOperationInfo"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:minHeight="?attr/actionBarSize"
+        android:theme="?attr/actionBarTheme" >
+
+        <TextView
+            android:id="@+id/textViewToolbar"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:text="아파트 운영정보"
+            android:textAppearance="@style/TextAppearance.AppCompat.Medium"
+            android:textStyle="bold" />
+
+    </com.google.android.material.appbar.MaterialToolbar>
+
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical"
+        android:padding="10dp">
+
+        <Button
+            android:id="@+id/buttonOperationInfo1"
+            style="?attr/materialButtonOutlinedStyle"
+            android:layout_width="match_parent"
+            android:layout_height="60dp"
+            android:layout_margin="10dp"
+            android:backgroundTint="#B8D7D5"
+            android:drawableLeft="@drawable/icon_operation_info1"
+            android:fontFamily="@font/notosans_semibold"
+            android:gravity="center"
+            android:includeFontPadding="false"
+            android:text="계약서 정보"
+            android:textColor="@color/black"
+            android:textSize="16sp"
+            app:iconTint="@color/black"
+            app:strokeColor="#000000"
+            app:strokeWidth="2dp"
+            tools:insetBottom="0dp"
+            tools:insetTop="0dp" />
+
+        <Button
+            android:id="@+id/buttonOperationInfo2"
+            style="?attr/materialButtonOutlinedStyle"
+            android:layout_width="match_parent"
+            android:layout_height="60dp"
+            android:layout_margin="10dp"
+            android:backgroundTint="#B8D7D5"
+            android:drawableLeft="@drawable/icon_operation_info2"
+            android:fontFamily="@font/notosans_semibold"
+            android:gravity="center"
+            android:includeFontPadding="false"
+            android:text="관리규약"
+            android:textColor="@color/black"
+            android:textSize="16sp"
+            app:iconTint="@color/black"
+            app:strokeColor="#000000"
+            app:strokeWidth="2dp"
+            tools:insetBottom="0dp"
+            tools:insetTop="0dp"/>
+
+        <Button
+            android:id="@+id/buttonOperationInfo3"
+            style="?attr/materialButtonOutlinedStyle"
+            android:layout_width="match_parent"
+            android:layout_height="60dp"
+            android:layout_margin="10dp"
+            android:backgroundTint="#B8D7D5"
+            android:drawableLeft="@drawable/icon_operation_info3"
+            android:fontFamily="@font/notosans_semibold"
+            android:gravity="center"
+            android:includeFontPadding="false"
+            android:text="입찰공고"
+            android:textColor="@color/black"
+            android:textSize="16sp"
+            app:iconTint="@color/black"
+            app:rippleColor="@color/third"
+            app:strokeColor="#000000"
+            app:strokeWidth="2dp"
+            tools:insetBottom="0dp"
+            tools:insetTop="0dp"/>
+
+        <Button
+            android:id="@+id/buttonOperationInfo4"
+            style="?attr/materialButtonOutlinedStyle"
+            android:layout_width="match_parent"
+            android:layout_height="60dp"
+            android:layout_margin="10dp"
+            android:backgroundTint="#B8D7D5"
+            android:drawableLeft="@drawable/icon_operation_info4"
+            android:fontFamily="@font/notosans_semibold"
+            android:gravity="center"
+            android:includeFontPadding="false"
+            android:text="부과명세서"
+            android:textColor="@color/black"
+            android:textSize="16sp"
+            app:iconTint="@color/black"
+            app:strokeColor="#000000"
+            app:strokeWidth="2dp"
+            tools:insetBottom="0dp"
+            tools:insetTop="0dp"/>
+
+        <Button
+            android:id="@+id/buttonOperationInfo5"
+            style="?attr/materialButtonOutlinedStyle"
+            android:layout_width="match_parent"
+            android:layout_height="60dp"
+            android:layout_margin="10dp"
+            android:backgroundTint="#B8D7D5"
+            android:drawableLeft="@drawable/icon_operation_info5"
+            android:fontFamily="@font/notosans_semibold"
+            android:gravity="center"
+            android:includeFontPadding="false"
+            android:text="제무제표"
+            android:textColor="@color/black"
+            android:textSize="16sp"
+            app:iconTint="@color/black"
+            app:strokeColor="#000000"
+            app:strokeWidth="2dp"
+            tools:insetBottom="0dp"
+            tools:insetTop="0dp"/>
+
+        <Button
+            android:id="@+id/buttonOperationInfo6"
+            style="?attr/materialButtonOutlinedStyle"
+            android:layout_width="match_parent"
+            android:layout_height="60dp"
+            android:layout_margin="10dp"
+            android:backgroundTint="#B8D7D5"
+            android:drawableLeft="@drawable/icon_operation_info6"
+            android:fontFamily="@font/notosans_semibold"
+            android:gravity="center"
+            android:includeFontPadding="false"
+            android:text="안전관리계획"
+            android:textColor="@color/black"
+            android:textSize="16sp"
+            app:iconTint="@color/black"
+            app:strokeColor="#000000"
+            app:strokeWidth="2dp"
+            tools:insetBottom="0dp"
+            tools:insetTop="0dp"/>
+
+        <Button
+            android:id="@+id/buttonOperationInfo7"
+            style="?attr/materialButtonOutlinedStyle"
+            android:layout_width="match_parent"
+            android:layout_height="60dp"
+            android:layout_margin="10dp"
+            android:backgroundTint="#B8D7D5"
+            android:drawableLeft="@drawable/icon_operation_info7"
+            android:fontFamily="@font/notosans_semibold"
+            android:gravity="center"
+            android:includeFontPadding="false"
+            android:text="수선계획"
+            android:textColor="@color/black"
+            android:textSize="16sp"
+            app:iconTint="@color/black"
+            app:strokeColor="#000000"
+            app:strokeWidth="2dp"
+            tools:insetBottom="0dp"
+            tools:insetTop="0dp"/>
+    </LinearLayout>
+
+</LinearLayout>
