+++ app/src/main/res/layout/fragment_survey_list.xml
@@ -60,7 +60,7 @@
                 android:layout_height="wrap_content">
 
                 <com.google.android.material.textfield.TextInputEditText
-                    android:id="@+id/surveyListTextViewWrite1"
+                    android:id="@+id/surveyListTextViewWrite"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:fontFamily="@font/notosans_semibold"
